requirements.txt
-i https://pypi.org/simple
asgiref==3.11.0; python_version >= '3.9'
certifi==2026.1.4; python_version >= '3.7'
charset-normalizer==3.4.4; python_version >= '3.7'
django==5.2.10; python_version >= '3.10'
django-cors-headers==4.9.0; python_version >= '3.9'
djangorestframework==3.16.1; python_version >= '3.9'
djangorestframework-simplejwt==5.5.1; python_version >= '3.9'
idna==3.11; python_version >= '3.8'
pyjwt==2.10.1; python_version >= '3.9'
requests==2.32.5; python_version >= '3.9'
sqlparse==0.5.5; python_version >= '3.8'
tzdata==2025.3; python_version >= '2'
urllib3==2.6.3; python_version >= '3.9'
